delete the old eulerian test cases
