--- tests/2d_examples/test_2d_eulerian_flow_around_cylinder/regression_test_tool/Cylinder_TotalViscousForceOnSolid.xml
@@ -1,19 +1,0 @@
-<?xml version="1.0" encoding="UTF-8" ?>
-<TotalViscousForceOnSolid>
-    <Snapshot_745 TotalViscousForceOnSolid_0="~[1.05865,-0.0137443]" />
-    <Snapshot_1516 TotalViscousForceOnSolid_0="~[1.0487,-0.0101599]" />
-    <Snapshot_2286 TotalViscousForceOnSolid_0="~[1.03734,-0.00766837]" />
-    <Snapshot_3055 TotalViscousForceOnSolid_0="~[1.03278,-0.0069651]" />
-    <Snapshot_3824 TotalViscousForceOnSolid_0="~[1.03259,-0.00694021]" />
-    <Snapshot_4593 TotalViscousForceOnSolid_0="~[1.03226,-0.00685662]" />
-    <Snapshot_5362 TotalViscousForceOnSolid_0="~[1.03216,-0.00683711]" />
-    <Snapshot_6131 TotalViscousForceOnSolid_0="~[1.03213,-0.00683118]" />
-    <Snapshot_6900 TotalViscousForceOnSolid_0="~[1.03212,-0.00682974]" />
-    <Snapshot_7669 TotalViscousForceOnSolid_0="~[1.03212,-0.00682894]" />
-    <Snapshot_8438 TotalViscousForceOnSolid_0="~[1.03212,-0.00682869]" />
-    <Snapshot_9207 TotalViscousForceOnSolid_0="~[1.03212,-0.00682864]" />
-    <Snapshot_9976 TotalViscousForceOnSolid_0="~[1.03212,-0.00682869]" />
-    <Snapshot_10745 TotalViscousForceOnSolid_0="~[1.03212,-0.00682867]" />
-    <Snapshot_11514 TotalViscousForceOnSolid_0="~[1.03212,-0.00682868]" />
-    <Snapshot_12283 TotalViscousForceOnSolid_0="~[1.03212,-0.0068287]" />
-</TotalViscousForceOnSolid>
