--- tests/2d_examples/test_2d_eulerian_flow_around_cylinder/regression_test_tool/Cylinder_TotalViscousForceOnSolid_Run_14_result.xml
@@ -1,9 +1,0 @@
-<?xml version="1.0" encoding="UTF-8" ?>
-<result>
-    <Snapshot_Element>
-        <Snapshot number_of_snapshot_for_local_result_="16" />
-    </Snapshot_Element>
-    <Result_Element>
-        <Particle_0 snapshot_0="~[1.00144,0.014526]" snapshot_1="~[0.987583,0.0224051]" snapshot_2="~[0.973087,0.0248384]" snapshot_3="~[0.967254,0.0270384]" snapshot_4="~[0.966568,0.0278639]" snapshot_5="~[0.966068,0.0281363]" snapshot_6="~[0.965899,0.0281989]" snapshot_7="~[0.965842,0.0282368]" snapshot_8="~[0.965822,0.0282479]" snapshot_9="~[0.965816,0.0282516]" snapshot_10="~[0.965814,0.0282528]" snapshot_11="~[0.965813,0.0282533]" snapshot_12="~[0.965813,0.0282535]" snapshot_13="~[0.965812,0.0282535]" snapshot_14="~[0.965812,0.0282535]" snapshot_15="~[0.965812,0.0282536]" />
-    </Result_Element>
-</result>
