--- tests/2d_examples/test_2d_eulerian_flow_around_cylinder/regression_test_tool/Cylinder_TotalViscousForceOnSolid_dtwdistance.xml
@@ -1,4 +1,0 @@
-<?xml version="1.0" encoding="UTF-8" ?>
-<dtw_distance>
-    <DTWDistance TotalViscousForceOnSolid_0="0.977749459334148" />
-</dtw_distance>
