--- tests/2d_examples/test_2d_eulerian_flow_around_cylinder/regression_test_tool/Cylinder_TotalViscousForceOnSolid_runtimes.dat
@@ -1,3 +1,0 @@
-true
-16
-4
