--- tests/2d_examples/test_2d_eulerian_taylor_green/CMakeLists.txt
@@ -1,26 +1,0 @@
-set(CMAKE_MODULE_PATH ${CMAKE_MODULE_PATH} ${SPHINXSYS_PROJECT_DIR}/cmake) # main (top) cmake dir
-
-set(CMAKE_VERBOSE_MAKEFILE on)
-
-STRING(REGEX REPLACE ".*/(.*)" "\\1" CURRENT_FOLDER ${CMAKE_CURRENT_SOURCE_DIR})
-PROJECT("${CURRENT_FOLDER}")
-
-SET(EXECUTABLE_OUTPUT_PATH "${PROJECT_BINARY_DIR}/bin/")
-SET(LIBRARY_OUTPUT_PATH ${PROJECT_BINARY_DIR}/lib)
-SET(BUILD_INPUT_PATH "${EXECUTABLE_OUTPUT_PATH}/input")
-SET(BUILD_RELOAD_PATH "${EXECUTABLE_OUTPUT_PATH}/reload")
-
-file(MAKE_DIRECTORY ${BUILD_INPUT_PATH})
-execute_process(COMMAND ${CMAKE_COMMAND} -E make_directory ${BUILD_INPUT_PATH})
-file(COPY ${CMAKE_CURRENT_SOURCE_DIR}/regression_test_tool/ DESTINATION ${BUILD_INPUT_PATH})
-
-aux_source_directory(. DIR_SRCS)
-ADD_EXECUTABLE(${PROJECT_NAME} ${DIR_SRCS})
-
-add_test(NAME ${PROJECT_NAME} COMMAND ${PROJECT_NAME} --state_recording=${TEST_STATE_RECORDING}
-    WORKING_DIRECTORY ${EXECUTABLE_OUTPUT_PATH})
-
-set_tests_properties(${PROJECT_NAME} PROPERTIES LABELS "periodic boundary, Eulerian")
-
-set_target_properties(${PROJECT_NAME} PROPERTIES VS_DEBUGGER_WORKING_DIRECTORY "${EXECUTABLE_OUTPUT_PATH}")
-target_link_libraries(${PROJECT_NAME} extra_sources_2d)
