--- tests/2d_examples/test_2d_eulerian_taylor_green/regression_test_tool/WaterBody_MaximumSpeed_ensemble_averaged_mean_variance.xml
@@ -1,107 +1,0 @@
-<?xml version="1.0" encoding="UTF-8" ?>
-<meanvariance>
-    <Mean_Element>
-        <Snapshot_141 MaximumSpeed_0="0.88607672273669535" />
-        <Snapshot_280 MaximumSpeed_0="0.79245451530510158" />
-        <Snapshot_418 MaximumSpeed_0="0.71145167039692436" />
-        <Snapshot_555 MaximumSpeed_0="0.64042982596343279" />
-        <Snapshot_692 MaximumSpeed_0="0.57780848176539401" />
-        <Snapshot_828 MaximumSpeed_0="0.52300825982976917" />
-        <Snapshot_963 MaximumSpeed_0="0.47450060808111344" />
-        <Snapshot_1098 MaximumSpeed_0="0.43094248174663569" />
-        <Snapshot_1232 MaximumSpeed_0="0.39222420244805339" />
-        <Snapshot_1366 MaximumSpeed_0="0.35741431478211466" />
-        <Snapshot_1499 MaximumSpeed_0="0.32622695091869552" />
-        <Snapshot_1632 MaximumSpeed_0="0.29796337546548507" />
-        <Snapshot_1764 MaximumSpeed_0="0.27256601342584369" />
-        <Snapshot_1896 MaximumSpeed_0="0.24952261187218683" />
-        <Snapshot_2028 MaximumSpeed_0="0.22856103912539169" />
-        <Snapshot_2160 MaximumSpeed_0="0.20947150540664475" />
-        <Snapshot_2291 MaximumSpeed_0="0.19220309838574134" />
-        <Snapshot_2422 MaximumSpeed_0="0.17645345653686292" />
-        <Snapshot_2553 MaximumSpeed_0="0.16206556691803306" />
-        <Snapshot_2684 MaximumSpeed_0="0.14890948978590487" />
-        <Snapshot_2815 MaximumSpeed_0="0.13687929476158661" />
-        <Snapshot_2945 MaximumSpeed_0="0.12592995865436768" />
-        <Snapshot_3075 MaximumSpeed_0="0.11589274968923528" />
-        <Snapshot_3205 MaximumSpeed_0="0.10668511552657363" />
-        <Snapshot_3335 MaximumSpeed_0="0.098238277273080446" />
-        <Snapshot_3465 MaximumSpeed_0="0.090536179604021924" />
-        <Snapshot_3595 MaximumSpeed_0="0.083456198194547582" />
-        <Snapshot_3725 MaximumSpeed_0="0.076945001167877392" />
-        <Snapshot_3855 MaximumSpeed_0="0.070952922553431122" />
-        <Snapshot_3985 MaximumSpeed_0="0.06543674723838952" />
-        <Snapshot_4115 MaximumSpeed_0="0.060358467012971241" />
-        <Snapshot_4244 MaximumSpeed_0="0.055725770134469325" />
-        <Snapshot_4373 MaximumSpeed_0="0.051452701278109414" />
-        <Snapshot_4502 MaximumSpeed_0="0.047510083100220361" />
-        <Snapshot_4631 MaximumSpeed_0="0.043872386505178246" />
-        <Snapshot_4760 MaximumSpeed_0="0.040515849126737349" />
-        <Snapshot_4889 MaximumSpeed_0="0.037417903353014002" />
-        <Snapshot_5018 MaximumSpeed_0="0.0345580866220839" />
-        <Snapshot_5147 MaximumSpeed_0="0.031918186024111984" />
-        <Snapshot_5276 MaximumSpeed_0="0.02948133047165202" />
-        <Snapshot_5405 MaximumSpeed_0="0.027231570647046736" />
-        <Snapshot_5534 MaximumSpeed_0="0.025154272291963282" />
-        <Snapshot_5663 MaximumSpeed_0="0.023236263557953479" />
-        <Snapshot_5792 MaximumSpeed_0="0.021465400398820281" />
-        <Snapshot_5921 MaximumSpeed_0="0.019830263112895392" />
-        <Snapshot_6050 MaximumSpeed_0="0.018320297027054937" />
-        <Snapshot_6179 MaximumSpeed_0="0.016925923025176717" />
-        <Snapshot_6308 MaximumSpeed_0="0.015638337734631965" />
-        <Snapshot_6437 MaximumSpeed_0="0.014452219524344436" />
-        <Snapshot_6566 MaximumSpeed_0="0.013361068890483543" />
-    </Mean_Element>
-    <Variance_Element>
-        <Snapshot_141 MaximumSpeed_0="7.8513195857580315e-05" />
-        <Snapshot_280 MaximumSpeed_0="6.2798415882744574e-05" />
-        <Snapshot_418 MaximumSpeed_0="5.0616347931057478e-05" />
-        <Snapshot_555 MaximumSpeed_0="4.1015036198355605e-05" />
-        <Snapshot_692 MaximumSpeed_0="3.338626416000322e-05" />
-        <Snapshot_828 MaximumSpeed_0="2.7353763985016448e-05" />
-        <Snapshot_963 MaximumSpeed_0="2.2515082706934689e-05" />
-        <Snapshot_1098 MaximumSpeed_0="1.857114225739498e-05" />
-        <Snapshot_1232 MaximumSpeed_0="1.5383982498601185e-05" />
-        <Snapshot_1366 MaximumSpeed_0="1.2774499241117019e-05" />
-        <Snapshot_1499 MaximumSpeed_0="1.0642402350570903e-05" />
-        <Snapshot_1632 MaximumSpeed_0="8.8782173118787431e-06" />
-        <Snapshot_1764 MaximumSpeed_0="7.4292231674857952e-06" />
-        <Snapshot_1896 MaximumSpeed_0="6.2261533835518578e-06" />
-        <Snapshot_2028 MaximumSpeed_0="5.2240148606079498e-06" />
-        <Snapshot_2160 MaximumSpeed_0="4.3878311577327326e-06" />
-        <Snapshot_2291 MaximumSpeed_0="3.6942031029079034e-06" />
-        <Snapshot_2422 MaximumSpeed_0="3.1135822323807329e-06" />
-        <Snapshot_2553 MaximumSpeed_0="2.6265247980463659e-06" />
-        <Snapshot_2684 MaximumSpeed_0="2.2174036148298572e-06" />
-        <Snapshot_2815 MaximumSpeed_0="1.8735941334429582e-06" />
-        <Snapshot_2945 MaximumSpeed_0="1.5858354486690755e-06" />
-        <Snapshot_3075 MaximumSpeed_0="1.3431129430532133e-06" />
-        <Snapshot_3205 MaximumSpeed_0="1.138171387491867e-06" />
-        <Snapshot_3335 MaximumSpeed_0="9.6507591215828944e-07" />
-        <Snapshot_3465 MaximumSpeed_0="8.1967998172917157e-07" />
-        <Snapshot_3595 MaximumSpeed_0="6.9649370170877445e-07" />
-        <Snapshot_3725 MaximumSpeed_0="5.9205332047248301e-07" />
-        <Snapshot_3855 MaximumSpeed_0="5.0343172188736908e-07" />
-        <Snapshot_3985 MaximumSpeed_0="4.2819678891408787e-07" />
-        <Snapshot_4115 MaximumSpeed_0="3.6431445401559379e-07" />
-        <Snapshot_4244 MaximumSpeed_0="3.1053614570798891e-07" />
-        <Snapshot_4373 MaximumSpeed_0="2.6473804688144555e-07" />
-        <Snapshot_4502 MaximumSpeed_0="2.2572079961900736e-07" />
-        <Snapshot_4631 MaximumSpeed_0="1.9247862976598779e-07" />
-        <Snapshot_4760 MaximumSpeed_0="1.6415340304608098e-07" />
-        <Snapshot_4889 MaximumSpeed_0="1.4000994913355891e-07" />
-        <Snapshot_5018 MaximumSpeed_0="1.1942613509796779e-07" />
-        <Snapshot_5147 MaximumSpeed_0="1.0187705990700094e-07" />
-        <Snapshot_5276 MaximumSpeed_0="8.6914884637899027e-08" />
-        <Snapshot_5405 MaximumSpeed_0="7.4155843990513531e-08" />
-        <Snapshot_5534 MaximumSpeed_0="6.3273741453827681e-08" />
-        <Snapshot_5663 MaximumSpeed_0="5.3992394413471375e-08" />
-        <Snapshot_5792 MaximumSpeed_0="4.607634142817413e-08" />
-        <Snapshot_5921 MaximumSpeed_0="3.9323933512676734e-08" />
-        <Snapshot_6050 MaximumSpeed_0="3.3563328315957106e-08" />
-        <Snapshot_6179 MaximumSpeed_0="2.8648687025422807e-08" />
-        <Snapshot_6308 MaximumSpeed_0="2.4455760710245276e-08" />
-        <Snapshot_6437 MaximumSpeed_0="2.0886664917986938e-08" />
-        <Snapshot_6566 MaximumSpeed_0="1.7851816189636245e-08" />
-    </Variance_Element>
-</meanvariance>
