--- tests/2d_examples/test_2d_eulerian_taylor_green/regression_test_tool/WaterBody_MaximumSpeed_runtimes.dat
@@ -1,3 +1,0 @@
-true
-6
-5
