--- tests/2d_examples/test_2d_eulerian_taylor_green/regression_test_tool/WaterBody_TotalMechanicalEnergy_ensemble_averaged_mean_variance.xml
@@ -1,109 +1,0 @@
-<?xml version="1.0" encoding="UTF-8" ?>
-<meanvariance>
-    <Mean_Element>
-        <Snapshot_0 TotalMechanicalEnergy_0="0.25000000000000006" />
-        <Snapshot_141 TotalMechanicalEnergy_0="0.19842357094638133" />
-        <Snapshot_280 TotalMechanicalEnergy_0="0.15890612609666291" />
-        <Snapshot_418 TotalMechanicalEnergy_0="0.12809529825685975" />
-        <Snapshot_555 TotalMechanicalEnergy_0="0.10386994502881187" />
-        <Snapshot_692 TotalMechanicalEnergy_0="0.084558700328274353" />
-        <Snapshot_828 TotalMechanicalEnergy_0="0.069186264867873229" />
-        <Snapshot_963 TotalMechanicalEnergy_0="0.056869476848208471" />
-        <Snapshot_1098 TotalMechanicalEnergy_0="0.04687617778248257" />
-        <Snapshot_1232 TotalMechanicalEnergy_0="0.038791956837521874" />
-        <Snapshot_1366 TotalMechanicalEnergy_0="0.032175512309050074" />
-        <Snapshot_1499 TotalMechanicalEnergy_0="0.02677910834337803" />
-        <Snapshot_1632 TotalMechanicalEnergy_0="0.022328935371416631" />
-        <Snapshot_1764 TotalMechanicalEnergy_0="0.018674063484721944" />
-        <Snapshot_1896 TotalMechanicalEnergy_0="0.015641161955121156" />
-        <Snapshot_2028 TotalMechanicalEnergy_0="0.013119101331616678" />
-        <Snapshot_2160 TotalMechanicalEnergy_0="0.011017758219784872" />
-        <Snapshot_2291 TotalMechanicalEnergy_0="0.0092751541110012384" />
-        <Snapshot_2422 TotalMechanicalEnergy_0="0.0078164414880587309" />
-        <Snapshot_2553 TotalMechanicalEnergy_0="0.0065935689930130895" />
-        <Snapshot_2684 TotalMechanicalEnergy_0="0.0055669884268849017" />
-        <Snapshot_2815 TotalMechanicalEnergy_0="0.0047040989596651506" />
-        <Snapshot_2945 TotalMechanicalEnergy_0="0.0039827782556991816" />
-        <Snapshot_3075 TotalMechanicalEnergy_0="0.0033744431023426664" />
-        <Snapshot_3205 TotalMechanicalEnergy_0="0.0028608760384061876" />
-        <Snapshot_3335 TotalMechanicalEnergy_0="0.0024269068285982611" />
-        <Snapshot_3465 TotalMechanicalEnergy_0="0.0020598810531163225" />
-        <Snapshot_3595 TotalMechanicalEnergy_0="0.0017492233387050307" />
-        <Snapshot_3725 TotalMechanicalEnergy_0="0.0014860832435710091" />
-        <Snapshot_3855 TotalMechanicalEnergy_0="0.0012630441668977264" />
-        <Snapshot_3985 TotalMechanicalEnergy_0="0.0010738809412370652" />
-        <Snapshot_4115 TotalMechanicalEnergy_0="0.00091336008821772598" />
-        <Snapshot_4244 TotalMechanicalEnergy_0="0.0007780969621312522" />
-        <Snapshot_4373 TotalMechanicalEnergy_0="0.00066305104974724203" />
-        <Snapshot_4502 TotalMechanicalEnergy_0="0.00056515437584343396" />
-        <Snapshot_4631 TotalMechanicalEnergy_0="0.00048181650682206667" />
-        <Snapshot_4760 TotalMechanicalEnergy_0="0.00041084729039722714" />
-        <Snapshot_4889 TotalMechanicalEnergy_0="0.00035039283327806112" />
-        <Snapshot_5018 TotalMechanicalEnergy_0="0.00029888203453030726" />
-        <Snapshot_5147 TotalMechanicalEnergy_0="0.00025498197621809008" />
-        <Snapshot_5276 TotalMechanicalEnergy_0="0.00021756078552818731" />
-        <Snapshot_5405 TotalMechanicalEnergy_0="0.00018565664200253878" />
-        <Snapshot_5534 TotalMechanicalEnergy_0="0.00015845173677347553" />
-        <Snapshot_5663 TotalMechanicalEnergy_0="0.00013525034892475918" />
-        <Snapshot_5792 TotalMechanicalEnergy_0="0.00011546038478154611" />
-        <Snapshot_5921 TotalMechanicalEnergy_0="9.8577803797612212e-05" />
-        <Snapshot_6050 TotalMechanicalEnergy_0="8.4173426094802842e-05" />
-        <Snapshot_6179 TotalMechanicalEnergy_0="7.1881739700530332e-05" />
-        <Snapshot_6308 TotalMechanicalEnergy_0="6.1391403886296148e-05" />
-        <Snapshot_6437 TotalMechanicalEnergy_0="5.2437186897731056e-05" />
-        <Snapshot_6566 TotalMechanicalEnergy_0="4.4793109924711826e-05" />
-    </Mean_Element>
-    <Variance_Element>
-        <Snapshot_0 TotalMechanicalEnergy_0="6.2500000000000028e-06" />
-        <Snapshot_141 TotalMechanicalEnergy_0="3.9371913507113674e-06" />
-        <Snapshot_280 TotalMechanicalEnergy_0="2.5251156911048547e-06" />
-        <Snapshot_418 TotalMechanicalEnergy_0="1.6408405435513886e-06" />
-        <Snapshot_555 TotalMechanicalEnergy_0="1.0788965480288433e-06" />
-        <Snapshot_692 TotalMechanicalEnergy_0="7.1501738012069211e-07" />
-        <Snapshot_828 TotalMechanicalEnergy_0="4.786739246367526e-07" />
-        <Snapshot_963 TotalMechanicalEnergy_0="3.2341373969889331e-07" />
-        <Snapshot_1098 TotalMechanicalEnergy_0="2.1973760434949191e-07" />
-        <Snapshot_1232 TotalMechanicalEnergy_0="1.5048159152841667e-07" />
-        <Snapshot_1366 TotalMechanicalEnergy_0="1.0352635923498388e-07" />
-        <Snapshot_1499 TotalMechanicalEnergy_0="7.1712064366638179e-08" />
-        <Snapshot_1632 TotalMechanicalEnergy_0="4.9858135482090372e-08" />
-        <Snapshot_1764 TotalMechanicalEnergy_0="3.4872064703142765e-08" />
-        <Snapshot_1896 TotalMechanicalEnergy_0="2.4464594730633075e-08" />
-        <Snapshot_2028 TotalMechanicalEnergy_0="1.7211081974922773e-08" />
-        <Snapshot_2160 TotalMechanicalEnergy_0="1.2139099618963829e-08" />
-        <Snapshot_2291 TotalMechanicalEnergy_0="8.6028483782824167e-09" />
-        <Snapshot_2422 TotalMechanicalEnergy_0="6.1096757536246474e-09" />
-        <Snapshot_2553 TotalMechanicalEnergy_0="4.3475152065624229e-09" />
-        <Snapshot_2684 TotalMechanicalEnergy_0="3.0991360145070949e-09" />
-        <Snapshot_2815 TotalMechanicalEnergy_0="2.2128547022323137e-09" />
-        <Snapshot_2945 TotalMechanicalEnergy_0="1.5862522634070504e-09" />
-        <Snapshot_3075 TotalMechanicalEnergy_0="1.1386866250948254e-09" />
-        <Snapshot_3205 TotalMechanicalEnergy_0="8.1846117071268643e-10" />
-        <Snapshot_3335 TotalMechanicalEnergy_0="5.8898767546970152e-10" />
-        <Snapshot_3465 TotalMechanicalEnergy_0="4.2431099529877199e-10" />
-        <Snapshot_3595 TotalMechanicalEnergy_0="3.0597822886704648e-10" />
-        <Snapshot_3725 TotalMechanicalEnergy_0="2.2084434068226013e-10" />
-        <Snapshot_3855 TotalMechanicalEnergy_0="1.5952805675344323e-10" />
-        <Snapshot_3985 TotalMechanicalEnergy_0="1.1532202759522554e-10" />
-        <Snapshot_4115 TotalMechanicalEnergy_0="8.3422665074913472e-11" />
-        <Snapshot_4244 TotalMechanicalEnergy_0="6.0543488247791519e-11" />
-        <Snapshot_4373 TotalMechanicalEnergy_0="4.3963669457094437e-11" />
-        <Snapshot_4502 TotalMechanicalEnergy_0="3.1939946853500113e-11" />
-        <Snapshot_4631 TotalMechanicalEnergy_0="2.3214714624623449e-11" />
-        <Snapshot_4760 TotalMechanicalEnergy_0="1.6879549602675447e-11" />
-        <Snapshot_4889 TotalMechanicalEnergy_0="1.2277513761263576e-11" />
-        <Snapshot_5018 TotalMechanicalEnergy_0="8.9330470564982498e-12" />
-        <Snapshot_5147 TotalMechanicalEnergy_0="6.5015808196087776e-12" />
-        <Snapshot_5276 TotalMechanicalEnergy_0="4.7332695399645647e-12" />
-        <Snapshot_5405 TotalMechanicalEnergy_0="3.4468388719662227e-12" />
-        <Snapshot_5534 TotalMechanicalEnergy_0="2.5106952886533538e-12" />
-        <Snapshot_5663 TotalMechanicalEnergy_0="1.8292656884271264e-12" />
-        <Snapshot_5792 TotalMechanicalEnergy_0="1.3331100453904429e-12" />
-        <Snapshot_5921 TotalMechanicalEnergy_0="9.7175834015619139e-13" />
-        <Snapshot_6050 TotalMechanicalEnergy_0="7.0851656605383744e-13" />
-        <Snapshot_6179 TotalMechanicalEnergy_0="5.1669845023756111e-13" />
-        <Snapshot_6308 TotalMechanicalEnergy_0="3.7689044711309824e-13" />
-        <Snapshot_6437 TotalMechanicalEnergy_0="2.7496585697480552e-13" />
-        <Snapshot_6566 TotalMechanicalEnergy_0="2.0064226967277656e-13" />
-    </Variance_Element>
-</meanvariance>
