+++ tests/user_examples/test_2d_eulerian_flow_around_cylinder_LG/CMakeLists.txt
@@ -20,7 +20,7 @@ add_test(NAME ${PROJECT_NAME}_particle_relaxation
      COMMAND ${PROJECT_NAME} --relax=true --state_recording=${TEST_STATE_RECORDING}
      WORKING_DIRECTORY ${EXECUTABLE_OUTPUT_PATH})
 add_test(NAME ${PROJECT_NAME}
-     COMMAND ${PROJECT_NAME}  --reload=true --state_recording=${TEST_STATE_RECORDING}
+     COMMAND ${PROJECT_NAME} --reload=true --state_recording=${TEST_STATE_RECORDING}
      WORKING_DIRECTORY ${EXECUTABLE_OUTPUT_PATH})
 set_tests_properties(${PROJECT_NAME} PROPERTIES DEPENDS "${PROJECT_NAME}_particle_relaxation")
-set_tests_properties(${PROJECT_NAME} PROPERTIES LABELS "periodic boundary, Eulerian")
+set_tests_properties(${PROJECT_NAME} PROPERTIES LABELS "Eulerian")
